+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Enh #91: Minor refactor internal class `RequestFactory`: explicitly mark read-only properties and add type to
   parameter of `create()` method (@vjik)
+- New #92: Add `EmitterInterface`, add `$emitter` parameter to `HttpApplicationRunner` constructor, and make
+  `SapiEmitter` public class (@vjik)
 
 ## 3.1.0 February 21, 2025