+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Yii Runner HTTP Change Log
 
+## 3.2.1 under development
+
+- no changes in this release.
+
 ## 3.2.0 June 14, 2025
 
 - New #90: Add `HttpApplicationRunner::runAndGetResponse()` method that returns response without emit (@vjik)